First 26 sec lap on the new speedway on the 2nd floor.

+++ stm32/main-board/2_Handler/Inc/motor.h
@@ -20,4 +20,4 @@
 
 void motorInit();
 void motorSetTorque(int16_t torqe);
-void motorSetDutyCycle(int8_t d);
+void motorSetDutyCycle(float d);
